// src/types/adminAnalytics.ts
export type AdminAnalyticsOverview = {
  activeUsers: number
  weeklyActiveStudents: number
  newRegistrations: number
  messagesSent: number
  filesUploaded: number
  teacherHelpRequests: number
  parentReportViews: number
  checkoutStarted: number
  checkoutCompleted: number
  cancelledSubscriptions: number
}